## Background
In the past, sonic use slice (`[]Node` or `[]Pair`) as fundamental storage of children nodes -- as user calls `Get()/Index()...` on one parent node, sonic will load its children nodes lazily (or so-called `on-demands`). But this design remains one problem --  once the underlying slice grows larger and moves to a new memory address, the children's address returned to users will be invalid and won't follow the change conducted on new address. A related issue #465  has been submit. Therefore, we must use a new kind of underlying storage to ensure the memory address returned will always be valid, and `linked list` looks like the only choice.
## Design
we designed a specific `linked chunk` for ast.Node, which stores every 8-nodes (or pairs) array as a chunk, and allocated one chunk by default. We provide `At(index)` and `Add(node)` two APIs to keep the same behavior with slice.  see details in `ast/buffer.go`. As for `Unset()`, now we just reset the node memory instead of removing its memory before, and specially ignore them when calling `MarshalJSON()`.  
- **Break Change**: for simplicity, we still count those nodes that have been unsettled. This means even if user conduct `Unset()` on one parent node,  its `Len()` and indexing of children will **NEVER** change.
## Performance 
By contrast, this design will reduce memory allocation when the slice size is less than 8 and gain better memory affinity. Meanwhile, we eliminate allocating memory for empty nodes. Therefore, most API related to memory allocation got improved.
```
name                            old time/op    new time/op     delta
LoadNode/InterfaceUseNode()-16    2.62µs ± 2%     2.19µs ± 4%  -16.42%  (p=0.000 n=8+10)
LoadNode/Load()-16                1.67µs ± 6%     1.30µs ± 4%  -22.47%  (p=0.000 n=10+10)
LoadNode/LoadAll()-16             7.06µs ± 6%     4.39µs ± 4%  -37.79%  (p=0.000 n=10+9)
LoadNode/Interface()-16           15.0µs ± 8%      9.1µs ±14%  -39.44%  (p=0.000 n=10+10)

ParseOne_Sonic-16      4.96µs ±11%     3.30µs ± 6%   -33.49%  (p=0.000 n=10+10)
ParseSeven_Sonic-16    10.5µs ± 8%      6.1µs ± 2%   -42.10%  (p=0.000 n=10+9)
ParseEmpty_Sonic-16    2.40µs ± 4%     0.82µs ± 8%   -66.06%  (p=0.000 n=10+9)

NodeGetByPath-16           14.0ns ±11%    14.9ns ±22%     ~     (p=0.237 n=8+10)
NodeIndex-16                    6.53ns ± 1%    5.53ns ± 3%  -15.39%  (p=0.000 n=8+10)
```
except sorting API will be a little slower since sonic's algothrim don't support linked-list
```
NodeSortKeys/recurse-16    6.55µs ± 2%    8.57µs ± 1%  +30.73%  (p=0.000 n=10+8)
NodeSortKeys/single-16     19.0ns ± 5%    20.6ns ± 3%   +8.39%  (p=0.000 n=9+10)

```